requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
qdrant-client>=1.7.0
openai>=1.10.0
httpx>=0.26.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
